# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/competitions": {
      "get": {
        "operationId": "CompetitionsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List competitions",
        "tags": [
          "Competitions"
        ]
      },
      "post": {
        "operationId": "CompetitionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompetitionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Create competition",
        "tags": [
          "Competitions"
        ]
      }
    },
    "/api/competitions/protocol-defaults/me": {
      "get": {
        "operationId": "CompetitionsController_getMyProtocolDefaults",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get my result protocol defaults",
        "tags": [
          "Competitions"
        ]
      },
      "patch": {
        "operationId": "CompetitionsController_saveMyProtocolDefaults",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveResultProtocolDefaultsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Save my result protocol defaults",
        "tags": [
          "Competitions"
        ]
      }
    },
    "/api/competitions/{id}": {
      "get": {
        "operationId": "CompetitionsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get competition by id",
        "tags": [
          "Competitions"
        ]
      },
      "patch": {
        "operationId": "CompetitionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompetitionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Update competition",
        "tags": [
          "Competitions"
        ]
      },
      "delete": {
        "operationId": "CompetitionsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Delete competition (admin only)",
        "tags": [
          "Competitions"
        ]
      }
    },
    "/api/competitions/{id}/age-groups": {
      "get": {
        "operationId": "CompetitionsController_getAgeGroups",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List age groups for competition",
        "tags": [
          "Competitions"
        ]
      },
      "post": {
        "operationId": "CompetitionsController_createAgeGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgeGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Create age group for competition",
        "tags": [
          "Competitions"
        ]
      }
    },
    "/api/competitions/{id}/result-protocol-config": {
      "get": {
        "operationId": "CompetitionsController_getResultProtocolConfigV2",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get result protocol configuration (v2)",
        "tags": [
          "Competitions"
        ]
      },
      "patch": {
        "description": "Configure protocol format, mixed groups, advanced grouping rules.",
        "operationId": "CompetitionsController_setResultProtocolConfigV2",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResultProtocolConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Set result protocol format for competition (v2)",
        "tags": [
          "Competitions"
        ]
      }
    },
    "/api/competitions/{id}/result-protocol-presets": {
      "get": {
        "operationId": "CompetitionsController_listResultProtocolPresets",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResultProtocolPresetResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "List result protocol presets for competition",
        "tags": [
          "Competitions"
        ]
      },
      "post": {
        "operationId": "CompetitionsController_createResultProtocolPreset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResultProtocolPresetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProtocolPresetResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Create result protocol preset",
        "tags": [
          "Competitions"
        ]
      }
    },
    "/api/competitions/{id}/result-protocol-presets/{presetId}": {
      "patch": {
        "operationId": "CompetitionsController_updateResultProtocolPreset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "presetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResultProtocolPresetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProtocolPresetResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Update result protocol preset",
        "tags": [
          "Competitions"
        ]
      },
      "delete": {
        "operationId": "CompetitionsController_deleteResultProtocolPreset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "presetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProtocolPresetResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Delete result protocol preset",
        "tags": [
          "Competitions"
        ]
      }
    },
    "/api/competitions/{id}/result-protocol-presets/{presetId}/apply": {
      "post": {
        "operationId": "CompetitionsController_applyResultProtocolPreset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "presetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetitionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Apply preset to competition protocol config",
        "tags": [
          "Competitions"
        ]
      }
    },
    "/api/competitions/age-groups/{agId}": {
      "delete": {
        "operationId": "CompetitionsController_deleteAgeGroup",
        "parameters": [
          {
            "name": "agId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Delete age group",
        "tags": [
          "Competitions"
        ]
      }
    },
...
  },
  "info": {
    "title": "SwimSync API",
    "description": "REST API for SwimSync services",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_token"
      }
    },
    "schemas": {
      "ResultProtocolConditionDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "enum": [
              "ageGroupId",
              "region",
              "club",
              "gender"
            ]
          },
          "operator": {
            "type": "string",
            "enum": [
              "EQ",
              "IN",
              "CONTAINS"
            ]
          },
          "value": {
            "description": "Condition value. For IN use array, for EQ/CONTAINS use scalar.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              {
                "type": "array",
                "items": {
                  "type": "number"
                }
              }
            ]
          }
        },
        "required": [
          "field",
          "operator",
          "value"
        ]
      },
      "ResultProtocolConditionBlockDto": {
        "type": "object",
        "properties": {
          "operator": {
            "type": "string",
            "enum": [
              "AND",
              "OR"
            ]
          },
          "conditions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResultProtocolConditionDto"
            }
          }
        },
        "required": [
          "operator",
          "conditions"
        ]
      },
      "ResultProtocolRuleGroupDto": {
        "type": "object",
        "properties": {
          "operator": {
            "type": "string",
            "enum": [
              "AND",
              "OR"
            ]
          },
          "conditions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResultProtocolConditionDto"
            }
          },
          "name": {
            "type": "string"
          },
          "subgroups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResultProtocolConditionBlockDto"
            }
          }
        },
        "required": [
          "operator",
          "conditions",
          "name"
        ]
      },
      "ResultProtocolAdvancedGroupingDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "default": false
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResultProtocolRuleGroupDto"
            }
          },
          "includeUnmatched": {
            "type": "boolean",
            "default": true
          },
          "unmatchedGroupName": {
            "type": "string",
            "default": "Інші"
          }
        },
        "required": [
          "enabled",
          "groups",
          "includeUnmatched"
        ]
      },
      "ResultProtocolConfigDto": {
        "type": "object",
        "properties": {
          "format": {
            "type": "string",
            "enum": [
              "SEPARATE",
              "COMBINED",
              "MIXED"
            ],
            "description": "Result protocol format"
          },
          "mixedFormatPrimaryAgeGroupId": {
            "type": "number",
            "nullable": true,
            "description": "Primary age group ID for MIXED format"
          },
          "mixedFormatSecondaryAgeGroupIds": {
            "description": "Secondary age group IDs for MIXED format",
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "advancedGrouping": {
            "nullable": true,
            "description": "Advanced custom grouping config (rules with AND/OR blocks, max 2 levels)",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ResultProtocolAdvancedGroupingDto"
              }
            ]
          }
        },
        "required": [
          "format"
        ]
      },
      "SaveResultProtocolDefaultsDto": {
        "type": "object",
        "properties": {
          "config": {
            "$ref": "#/components/schemas/ResultProtocolConfigDto"
          }
        },
        "required": [
          "config"
        ]
      },
      "CreateCompetitionDto": {
        "type": "object",
        "properties": {}
      },
      "CreateAgeGroupDto": {
        "type": "object",
        "properties": {}
      },
      "ResultProtocolPresetResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "competitionId": {
            "type": "number"
          },
          "ownerUserId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "isShared": {
            "type": "boolean"
          },
          "configJson": {
            "$ref": "#/components/schemas/ResultProtocolConfigDto"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "competitionId",
          "ownerUserId",
          "name",
          "isShared",
          "configJson",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateResultProtocolPresetDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "isShared": {
            "type": "boolean",
            "default": false
          },
          "config": {
            "$ref": "#/components/schemas/ResultProtocolConfigDto"
          }
        },
        "required": [
          "name",
          "config"
        ]
      },
      "UpdateResultProtocolPresetDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "isShared": {
            "type": "boolean"
          },
          "config": {
            "$ref": "#/components/schemas/ResultProtocolConfigDto"
          }
        }
      },
      "CompetitionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "categoriesStr": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "venue": {
            "type": "string"
          },
          "poolLength": {
            "type": "number"
          },
          "lanes": {
            "type": "number"
          },
          "dateFrom": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "dateTo": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "active",
              "completed"
            ]
          },
          "circularSeeding": {
            "type": "boolean"
          },
          "resultProtocolFormat": {
            "type": "string",
            "enum": [
              "SEPARATE",
              "COMBINED",
              "MIXED"
            ]
          },
          "mixedFormatPrimaryAgeGroupId": {
            "type": "number",
            "nullable": true
          },
          "mixedFormatSecondaryAgeGroupIds": {
            "type": "string",
            "nullable": true
          },
          "resultProtocolConfigJson": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ResultProtocolConfigDto"
              }
            ]
          },
          "createdByUserId": {
            "type": "number",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "categoriesStr",
          "location",
          "venue",
          "poolLength",
          "lanes",
          "status",
          "circularSeeding",
          "resultProtocolFormat",
          "createdAt"
        ]
      },
      "UpdateCompetitionDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}